Issue 2: Add parameter for the number of cores
Resolves issue 2. Added a command line option --threads which
allows the user to specify the number of concurrent threads to
use in the SimpleRunner. Note that this will probably be ignored
for any other runner we implement.

Also fixes the default to be cpu_count/2. This was the original
intention so that the experiment doesn't monopolize all the
resources on the machine.